Add functionality to determine if item is popular in scope of request; add unit test for ItemMappingService.isItemPopular(); refactor code: remove redundant classes, rename files

+++ items-data/src/main/java/com/ecommerce/itemsdata/repository/ItemDetailsRepository.java
@@ -4,8 +4,7 @@ import com.ecommerce.itemsdata.model.ItemDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
-public interface ItemDetailsRepository extends JpaRepository<ItemDetails, Integer> {
+public interface ItemDetailsRepository extends JpaRepository<ItemDetails, Long> {
+
 }
